C++ compat: Add type casts for `BMIN()` and `BMAX()`.

+++ maple/window.c
@@ -422,7 +422,7 @@ pmsg2_body(const char *msg)
 
     grayout(0, b_lines, GRAYOUT_DARK);
 
-    len = BMAX(strlen(msg), 16);  /* 取 msg title 其中較長者為 len */
+    len = BMAX(strlen(msg), (size_t)16); /* 取 msg title 其中較長者為 len */
     if (len % 2)                /* 變成偶數 */
         len++;
     x = (b_lines - 4) >> 1;     /* 置中 */
