+++ util/account.c
@@ -232,7 +232,7 @@ draw_vote(
         qsort(choice, items, sizeof(struct Tchoice), (int (*)(const void *lhs, const void *rhs))TchoiceCompare);
 
     if (vch->vpercent == '%')
-        fd = BMAX(1, total / sizeof(int));
+        fd = BMAX((size_t)1, total / sizeof(int));
     else
         fd = 0;
